Rename file, for consistency reasons only
