+++ compass-js-frontend/src/main/webapp/model/compass-model.js
