+++ compass-js-frontend/src/main/webapp/model/scenario.js
@@ -7,7 +7,7 @@
 
 "use strict";
 
-var CompassModel = require("./compassmodel");
+var CompassModel = require("./compass-model");
 var Config = require("../config");
 
 var Scenario = CompassModel.extend({
